HSEARCH-2841 Move ElasticsearchScrollAPIDocumentExtractor to a separate file
